pyproject.toml
[project]
name = "datadog-21st"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "braintrust>=0.5.6",
    "python-dotenv>=1.2.1",
    "supabase>=2.28.0",
]
